mips: Move MIPS boards into their own files
Since the I2000 (Rx2030) and R3030 (Rx3230) boards are so different,
they really belong in their own files. I've kept the original mips.cpp
file for common functions.

There doesn't seem to be enough commonality at this time to make a
common base class worthwhile, perhaps as more boards are implemented
some will be found.
