+++ src/mame/mips/mips.h
@@ -1,0 +1,8 @@
+#ifndef MAME_MIPS_MIPS_H
+
+#pragma once
+
+extern void mips_scsi_devices(device_slot_interface &device);
+extern void mips_rs232_devices(device_slot_interface &device);
+
+#endif // MAME_MIPS_MIPS_H
